# CLN/PERF: avoid double-checks