--- a/pandas/core/frame.py
+++ b/pandas/core/frame.py
@@ -3867,9 +3867,13 @@ class DataFrame(NDFrame, OpsMixin):
 
             series = self._get_item_cache(col)
             loc = self.index.get_loc(index)
-            if not can_hold_element(series._values, value):
-                # We'll go through loc and end up casting.
-                raise TypeError
+            dtype = series.dtype
+            if isinstance(dtype, np.dtype) and dtype.kind not in ["m", "M"]:
+                # otherwise we have EA values, and this check will be done
+                #  via setitem_inplace
+                if not can_hold_element(series._values, value):
+                    # We'll go through loc and end up casting.
+                    raise TypeError
 
             series._mgr.setitem_inplace(loc, value)
             # Note: trying to use series._set_value breaks tests in