--- a/pandas/core/internals/blocks.py
+++ b/pandas/core/internals/blocks.py
@@ -356,7 +356,7 @@ class Block(PandasObject):
     def dtype(self) -> DtypeObj:
         return self.values.dtype
 
-    def iget(self, i: int | tuple[int, int] | tuple[Literal[slice(None), int]]):
+    def iget(self, i: int | tuple[int, int] | tuple[Literal[slice(None)], int]):
         # Note: only reached with self.ndim == 2
         return self.values[i]
 
@@ -1483,11 +1483,11 @@ class ExtensionBlock(libinternals.Block, EABackedBlock):
             return (len(self.values),)
         return len(self._mgr_locs), len(self.values)
 
-    def iget(self, col: int | tuple[int, int] | tuple[Literal[slice(None), int]]):
+    def iget(self, col: int | tuple[int, int] | tuple[Literal[slice(None)], int]):
         # Note: only reached with self.ndim == 2
         # We _could_ make the annotation more specific, but mypy would#
         #  complain about override mismatch:
-        #  Literal[0] | tuple[Literal[0], int] | tuple[Literal[slice(None), int]]
+        #  Literal[0] | tuple[Literal[0], int] | tuple[Literal[slice(None)], int]
 
         if isinstance(col, tuple):
             # TODO(EA2D): unnecessary with 2D EAs