@jreback 2022-01-08
@meeseeksdev backport 1.4.x